+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # master
 
-- Removed Response struct, added Output (@BenchR267)
+- Removed Response struct, added package wide result handling (@BenchR267)
 - Add tests (@kiliankoe)
 - Added linting phase to danger (@BenchR267)
 - Add support for re-normalizing user input (@kiliankoe)